Sync

+++ .shell/zfs.sh
@@ -1,4 +1,9 @@
 #!/bin/sh
 
-time dd if=/dev/urandom of=/data/test.img bs=1M count=100 conv=fsync
-rm -f /data/test.img
+if [ "$2" == "0" ];
+    then
+        time dd if=/dev/urandom of=/data/test.img bs=1M count=$1 conv=fsync &
+    else
+        time dd if=/dev/urandom of=/data/test.img bs=1M count=$1 conv=fsync
+        rm -f /data/test.img
+fi
